requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
Flask-API==1.1
Flask-PyMongo==2.2.0
funcsigs==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mock==2.0.0
more-itertools==5.0.0
passlib==1.7.1
pathlib2==2.3.3
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
pycrypto==2.6.1
PyJWT==1.7.1
pymongo==3.7.2
pytest==4.3.1
scandir==1.10.0
six==1.12.0
Werkzeug==0.14.1
